Fixed a bug, caused another
Fixed a bug where adding a new card would not show in the list right
away and the list would be messed up.
The fix caused the app to always prompt the user to save even when the
user didn't make any changes, which means the TextWatcher isn't really
working properly.